Remove collapsed cell struct everywhere
