+++ src/generation/object/lib/mod.rs
@@ -1,13 +1,13 @@
 mod cell;
-mod collapsed_cell;
 mod connection_type;
+mod object_data;
 mod object_grid;
 mod object_name;
 mod wfc_status;
 
 pub use cell::Cell;
-pub use collapsed_cell::{CollapsedCell, ObjectData};
 pub use connection_type::Connection;
+pub use object_data::ObjectData;
 pub use object_grid::ObjectGrid;
 pub use object_name::ObjectName;
 pub use wfc_status::IterationResult;
